Removed unecessary guesstimate_form index file, deleted obsolete other files
