+++ src/components/distributions/summary/style.css
@@ -2,6 +2,7 @@
 
 .DistributionSummary {
   font-size: 1.6em;
+  font-weight: 500;
 }
 
 .DistributionSummary .mean {
