--- src/components/metrics/card/editing_pane/style.css
@@ -1,88 +1,0 @@
-@import './styles/variables.css';
-
-.GuesstimateForm input {
-  @mixin hover-input;
-}
-
-.GuesstimateForm.large input {
-  font-size: 1.5em;
-  border-radius: 3px;
-  font-weight: 100;
-  color: #666;
-  width: 87%;
-}
-
-.GuesstimateForm input{
-  float: left;
-  width: 75%;
-}
-
-.DistributionSelector {
-  width: 100%;
-}
-
-.DistributionSelector .DistributionList{
-  float: right;
-}
-
-.DistributionSelector hr{
-  border: none;
-  border-top: 1px solid $grey-1;
-}
-
-.GuesstimateForm .DistributionIcon.DistributionIcon{
-  font-size: 1em;
-  padding-top: 2px;
-  padding-left: 2px;
-  padding-right: 2px;
-  position: relative;
-  height: 26px;
-  width: 26px;
-  padding-top: 0;
-}
-
-.GuesstimateForm.large .DistributionIcon.DistributionIcon {
-  height: 34px;
-  width: 34px;
-}
-
-.DistributionIcon.button.selected{
-  background-color: $blue-2;
-}
-
-.DistributionIcon img{
-  width: 26px;
-  opacity: 1;
-  position: absolute;
-  left: 0;
-  right: 0;
-  top: 4px;
-  bottom: 0;
-}
-
-.GuesstimateForm.large .DistributionIcon img{
-  width: 34px;
-}
-
-.DistributionIcon.DistributionSelectorToggle{
-  display: inline-block;
-  color: #999;
-  float: right;
-}
-
-.DistributionIcon.DistributionSelectorToggle {
-  opacity: 0.6;
-}
-
-.button.DistributionIcon.DistributionSelectorToggle {
-  cursor: pointer;
-  border-radius: 5px;
-}
-
-.button.DistributionIcon.DistributionSelectorToggle:hover {
-  opacity: 1;
-}
-
-.button.DistributionIcon.DistributionSelectorToggle:hover {
-  background-color: $grey-1;
-}
