+++ src/components/simulations/histogram/style.css
@@ -7,8 +7,9 @@
 }
 
 .ReactModal__Overlay--after-open .react-d3-histogram__tick text {
-  fill: #aaa;
-  font-size: 12px;
+  fill: #666;
+  font-size: 13px;
+  font-weight: 600;
 }
 
 .react-d3-histogram svg {
